Convert app to work in wordpress
Converted the app to be a plugin.
Added the form to be a shortcode.
Modified the upload paths so they are relative to the server.
Updated the Link URL for the image that goes in the PDF to be a site URL
not a server directory.

Current Status:
Shortcode calls in forms with no errors.
Upload is not being put in the plugins uploads.
PDF is not being generated.